requirements.txt
django==4.2.7
djangorestframework==3.14.0
python-dotenv==1.0.0
coverage==7.3.2
django-redis==5.4.0
gunicorn==21.2.0
psycopg2-binary==2.9.9
celery[redis]==5.2.7
celery==5.2.7
Pillow==10.2.0